Updates to support constructing internal triangulation for dataset
Data like the NAM model is not on a regular structured grid, it is
on a curvilinear grid in geographic coordinates. Therefore, it is
necessary to handle its interpolation differently. This work
constructs a triangulation for that case using a constrained
triangulation generator and stores it as a LinearTriInterpolator
object, which is re-used for subsequent files, so the generation
is not performed for each file