Writing triangular interpolator
The triangular interpolator from Matplotlib was very slow
and memory intensive for this use case. It was attempting to use
approx 3GB of memory for the HRRR-conus model, which caused errors
within Argo (OOM). Instead, writing our own interpolator which
allows us to save off the interpolation weights and re-use. For the
same HRRR data, only 1.1GB of memory is used.